Tilia chron fixes

+++ function/ti/getanalysisunitbyid.sql
@@ -1,11 +1,14 @@
 CREATE OR REPLACE FUNCTION ti.getanalysisunitbyid(_analunitid integer)
-RETURNS TABLE(collectionunitid integer, analysisunitname character varying, depth numeric, thickness numeric)
+RETURNS TABLE(collectionunitid integer,
+              analysisunitname character varying,
+              depth double precision,
+              thickness double precision)
 LANGUAGE sql
 AS $function$
   SELECT au.collectionunitid,
          au.analysisunitname,
-         au.depth::numeric,
-         au.thickness::numeric
+         au.depth,
+         au.thickness
   FROM ndb.analysisunits AS au
   WHERE analysisunitid = _analunitid;
 $function$
